Reverted: mono command line reading as it broke the scheduler and at the moment provides me little benefit
Added: totalMegaBytesOfMemoryUsed which reads the total memory usage of that eve instance in MegaBytes
Added: a character specific setting EVEProcessMemoryCieling that needs to be specified in MegaBytes (ex. 950)
When totalMegaBytesOfMemoryUsed is above EVEProcessMemoryCieling between missions EVE is closed. This is a better alternative to EVE using more and more memory until EVE grinds to a halt from all the paging to disk. Keep in mind their are various ways to restart Questor periodically.

Added: tracking of the wallet balance. If the wallet balance is not updated in x min (setting: walletbalancechangelogoffdelay) that instance of eve is closed as questor is obviously "stuck" somehow. keep in mind there are various ways to restart questor periodically.

Added: statistics for lowest shield / armor / capacitor per mission and per pocket of a mission
Changed: the states around a bit to make restarting during a storyline and other "odd" places recover more gracefully. Questor will now *always* gotobase on restart it will also *always* unloadloot on restart, thus avoiding the "double ammo" situation that could occur before these changes.

Added: NotEnoughDrones and NotEnoughAmmo state awareness and recovery if someone manually corrects the issue, no restart of questor needed
Added: WarpOutStation state: if the bookmark exists it will be used, if it does not exist this will be skipped
Added: some debigging states that can only be activated via the questor gui