Added: bookmark sharing using the corp hangar
Added: MoveMissionCompletionItems with the intent that eventually this will move all mission completion items to the local hangar. Think: "marines" that you loot on a mission that does not require them for completion. You could then use them automagically for another mission because they are located in the right place rather than mixed in with your loot.