missed these with the questor.cs updates
this needs to be tested. As it stands this is likely non-functional and may get you flagged (thus killed)

This code is NOT used by questor in normal operation. You need to select one of the scoop states from the GUI to activate this behavior, yes, this needs a bit of cleanup.. patches welcomed