Enable the wreck blacklist feature since ppl seem to be adventurous and doing so on their own...
Created settings in the character xml for Blacklisting small and medium wrecks.
Keep in mind that these settings do currently seem to be incompatible with the noctis and should only be enabled for debugging purposes until deemed stable.

by making this commit those that have enabled the feature can disable it or enable it on a per character basis