Add a setting in your characters setting xml to either exit or logoff when hitting the EVEprocessMemoryCeiling or the WalletBalanceChangeLogoffDelay
I recommend this be kept at the default of exit. If you need to relaunch your Questor Sessions automatically you can use isboxer and a scheduled task.
The logoff option in theory works fine, in practice I had clients that would look closed but the eve process kept running which blocked isboxer from restarting that session (as far as it could tell that session WAS still running).