Cleanup of #compile#.bat
it is not more configurable. If you need to change the location of msbuild you can do that in the top of the script.
If later we need to add more programs to be compiled it is more modular.

You can optionally change the compile type from Debug to Release
*every* build will do a clean before building.