Add parameter to clearpocket called: NoMovement
If NoMovement is added as a parameter to clearpocket, clearpocket will issue NO Movement commands
This will need testing but it basically only targets things inside weapons range so it WILL NOT clear the whole pocket, only clear what is inside weapons range...
So you could use MoveToBackground then clearpocket NoMovement in an XML and be able to kill while moving to the next gate... you WILL need to issue another clearpocket with no paramaters to make sure the pocket is indeed clear.

if you find any issues with this action please let me know.