ADD: OptimalRange setting for the non-speedtanked gunboats so that they will stop abusing the orbit settings that the speedtanks use, which generally causes havoc
You can now specify your optimalrange if not speed tanking and you will approach until in optimal.
If you do not specify optimalrange in your config your weapons range is used to determine where you stop approaching.
The logs generated are quite clear on when and why your ship is stopping. Please read the logs. The defaults should be fine for everyone.

This only fixes clearpocket, more changes to come to the other actions