FIX: long standing lag issue when questor went into the idle state...
even figuring out that it WAS the idle state causing the problem was a bear.

The after mission logging was polling for the mission name in an inefficient way every pulse.
This has now been resolved, it only runs ONCE per mission.

I think don't think this was a problem until a few patches ago, but it is resolved now.